render_requirements.txt
beautifulsoup4==4.12.3
email-validator==2.1.1
flask==3.0.3
flask-sqlalchemy==3.1.1
google-genai==0.8.2
gunicorn==23.0.0
pillow==10.4.0
pocketsphinx==5.0.3
psycopg2-binary==2.9.9
pyrogram==2.0.106
requests==2.32.3
sift-stack-py==1.2.1
speechrecognition==3.10.4
sqlalchemy==2.0.35
trafilatura==1.12.2
werkzeug==3.0.4